Refactor package management
