+++ bash/.bashrc
@@ -1,0 +1,8 @@
+cat << EndOfMessage
+    ____     ______    __  ___    __  ___    ______
+   / __ \   / ____/   /  |/  /   /  |/  /   / ____/
+  / /_/ /  / __/     / /|_/ /   / /|_/ /   / __/
+ / _, _/  / /___    / /  / /   / /  / /   / /___
+/_/ |_|  /_____/   /_/  /_/   /_/  /_/   /_____/
+Distributed Public Key Infrastructure (PKI) protocol
+EndOfMessage
