+++ remme/__main__.py
@@ -18,8 +18,8 @@ import sys
 import logging
 import argparse
 from sawtooth_sdk.processor.core import TransactionProcessor
-from processor.certificate.certificate_handler import CertificateHandler
-from processor.token.token_handler import TokenHandler
+from remme.certificate.certificate_handler import CertificateHandler
+from remme.token.token_handler import TokenHandler
 
 LOGGER = logging.getLogger(__name__)
