+++ remme/shared/exceptions.py
