+++ setup.py
@@ -2,10 +2,10 @@ from setuptools import setup, find_packages
 
 data_files = []
 setup(
-    name='token',
+    name='remme',
     version='0.1',
-    description='REMME Token cli',
-    author='REMME Team',
+    description='REMME',
+    author='REMME',
     url='',
     packages=find_packages(),
     install_requires=[
@@ -13,8 +13,8 @@ setup(
     data_files=data_files,
     entry_points={
         'console_scripts': [
-            'rem-token=processor.token.token_cli:main',
-            'rem-crt=processor.certificate.certificate_cli:main'
+            'rem-token=remme.token.token_cli:main',
+            'rem-crt=remme.certificate.certificate_cli:main'
         ]
     }
 )
